Add Plant model template
The Plant model takes a long time to cover the attractor, so it would be beneficial to run a lot of separate trajectories and cluster them together. I haven't had luck with template discovery as-is, but extending the Template module to handle multiple trajectories should help to make this tractable.